+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+
+
+## [2.0.0] - 2017-03-17
 ### Added
 - Able to download audio and video content too
 
@@ -34,5 +37,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Getting thumbnail for ID3 album art
 - Session based download cache
 
-[unreleased]: https://github.com/andras-tim/tiatube/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/andras-tim/tiatube/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/andras-tim/tiatube/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/andras-tim/tiatube/compare/v1.0.0...v1.1.0